+++ Engine/JocysCom/Configuration/SettingsHelper.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using System.IO.Compression;
 using System.Linq;
@@ -151,6 +152,45 @@ namespace JocysCom.ClassLibrary.Configuration
 			for (int i = 0; i < bytes.Length; ++i)
 				crc = (crc >> 8) ^ table[(byte)(((crc) & 0xff) ^ bytes[i])];
 			return ~crc;
+		}
+
+		#endregion
+
+		#region  Synchronizing
+
+		/// <summary>
+		/// Synchronize source collection to destination.
+		/// </summary>
+		/// <remarks>
+		/// Same Code:
+		/// JocysCom\Controls\SearchHelper.cs
+		/// JocysCom\Configuration\SettingsHelper.cs
+		/// </remarks>
+		public static void Synchronize<T>(IList<T> source, IList<T> target)
+		{
+			// Convert to array to avoid modification of collection during processing.
+			var sList = source.ToArray();
+			var t = 0;
+			for (var s = 0; s < sList.Length; s++)
+			{
+				var item = sList[s];
+				// If item exists in destination and is in the correct position then continue
+				if (t < target.Count && target[t].Equals(item))
+				{
+					t++;
+					continue;
+				}
+				// If item is in destination but not at the correct position, remove it.
+				var indexInDestination = target.IndexOf(item);
+				if (indexInDestination != -1)
+					target.RemoveAt(indexInDestination);
+				// Insert item at the correct position.
+				target.Insert(s, item);
+				t = s + 1;
+			}
+			// Remove extra items.
+			while (target.Count > sList.Length)
+				target.RemoveAt(target.Count - 1);
 		}
 
 		#endregion