+++ Engine/JocysCom/Controls/ControlsHelper.WPF.cs
@@ -636,7 +636,13 @@ namespace JocysCom.ClassLibrary.Controls
 			return list;
 		}
 
+		[Obsolete]
 		public static void RestoreSelection<T>(DataGrid grid, string keyPropertyName, List<T> list, bool selectFirst = true)
+		{
+			RestoreSelection(grid, keyPropertyName, list, selectFirst ? 0 : -1);
+		}
+
+		public static void RestoreSelection<T>(DataGrid grid, string keyPropertyName, List<T> list, int selectIndex = 0)
 		{
 			if (grid is null)
 				throw new ArgumentNullException(nameof(grid));
@@ -673,9 +679,9 @@ namespace JocysCom.ClassLibrary.Controls
 						grid.SelectedItems.Add(item);
 				}
 			}
-			// If must select first row and nothing is selected then...
-			if (selectFirst && grid.SelectedItems.Count == 0)
-				grid.SelectedItem = items[0];
+			// If nothing was selected and must select index then...
+			if (grid.SelectedItems.Count == 0 && selectIndex >= 0 && selectIndex < grid.Items.Count)
+				grid.SelectedItem = items[selectIndex];
 		}
 
 		#endregion